requirements.txt
django==1.6
south
gunicorn==18.0
psycopg2==2.5.1
django-bootstrap3==2.5.6
django-markupfield==1.2.0
django-mptt==0.6.0
django-mptt-tree-editor==0.1.1
